requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# PDF Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0
pymupdf>=1.23.0
Pillow>=10.0.0
reportlab>=4.0.0  # For PDF overlay rendering

# Azure Services (Core only)
azure-ai-formrecognizer>=3.3.0  # Azure Document Intelligence
azure-storage-blob>=12.19.0  # Azure Blob Storage (optional)
azure-identity>=1.15.0  # Azure Managed Identity

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6  # For file uploads

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
aiosqlite>=0.19.0  # Async SQLite for local development
pyodbc>=5.0.0  # Azure SQL Database (optional, for production)

# Data Processing
pandas>=2.1.0
openpyxl>=3.1.0  # Excel export

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
requests>=2.31.0  # For demo scripts
openai>=1.13.0  # Azure OpenAI client
jsonschema>=4.22.0  # JSON Schema validation

# Streamlit HITL UI
streamlit>=1.28.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0  # For FastAPI TestClient (async)

# Development (optional)
black>=23.10.0
flake8>=6.1.0

